binder: fix c++11 narrowing violations
Summary:
The best fix for these narrowing
violations is to use static casting

Test Plan: verified

Reviewers: randall.rushing, synergy

Reviewed By: synergy

Differential Revision: http://review.codefirex.com/D85

+++ libs/binder/IPCThreadState.cpp
@@ -740,7 +740,7 @@ status_t IPCThreadState::waitForResponse(Parcel *reply, status_t *acquireResult)
             err = FAILED_TRANSACTION;
             goto finish;
         
-        case BR_ACQUIRE_RESULT:
+        case static_cast<int32_t>(BR_ACQUIRE_RESULT):
             {
                 ALOG_ASSERT(acquireResult != NULL, "Unexpected brACQUIRE_RESULT");
                 const int32_t result = mIn.readInt32();
@@ -749,7 +749,7 @@ status_t IPCThreadState::waitForResponse(Parcel *reply, status_t *acquireResult)
             }
             goto finish;
         
-        case BR_REPLY:
+        case static_cast<int32_t>(BR_REPLY):
             {
                 binder_transaction_data tr;
                 err = mIn.read(&tr, sizeof(tr));
@@ -950,14 +950,14 @@ status_t IPCThreadState::executeCommand(int32_t cmd)
     status_t result = NO_ERROR;
     
     switch (cmd) {
-    case BR_ERROR:
+    case static_cast<int32_t>(BR_ERROR):
         result = mIn.readInt32();
         break;
         
     case BR_OK:
         break;
         
-    case BR_ACQUIRE:
+    case static_cast<int32_t>(BR_ACQUIRE):
         refs = (RefBase::weakref_type*)mIn.readInt32();
         obj = (BBinder*)mIn.readInt32();
         ALOG_ASSERT(refs->refBase() == obj,
@@ -973,7 +973,7 @@ status_t IPCThreadState::executeCommand(int32_t cmd)
         mOut.writeInt32((int32_t)obj);
         break;
         
-    case BR_RELEASE:
+    case static_cast<int32_t>(BR_RELEASE):
         refs = (RefBase::weakref_type*)mIn.readInt32();
         obj = (BBinder*)mIn.readInt32();
         ALOG_ASSERT(refs->refBase() == obj,
@@ -986,7 +986,7 @@ status_t IPCThreadState::executeCommand(int32_t cmd)
         mPendingStrongDerefs.push(obj);
         break;
         
-    case BR_INCREFS:
+    case static_cast<int32_t>(BR_INCREFS):
         refs = (RefBase::weakref_type*)mIn.readInt32();
         obj = (BBinder*)mIn.readInt32();
         refs->incWeak(mProcess.get());
@@ -995,7 +995,7 @@ status_t IPCThreadState::executeCommand(int32_t cmd)
         mOut.writeInt32((int32_t)obj);
         break;
         
-    case BR_DECREFS:
+    case static_cast<int32_t>(BR_DECREFS):
         refs = (RefBase::weakref_type*)mIn.readInt32();
         obj = (BBinder*)mIn.readInt32();
         // NOTE: This assertion is not valid, because the object may no
@@ -1007,7 +1007,7 @@ status_t IPCThreadState::executeCommand(int32_t cmd)
         mPendingWeakDerefs.push(refs);
         break;
         
-    case BR_ATTEMPT_ACQUIRE:
+    case static_cast<int32_t>(BR_ATTEMPT_ACQUIRE):
         refs = (RefBase::weakref_type*)mIn.readInt32();
         obj = (BBinder*)mIn.readInt32();
          
@@ -1022,7 +1022,7 @@ status_t IPCThreadState::executeCommand(int32_t cmd)
         }
         break;
     
-    case BR_TRANSACTION:
+    case static_cast<int32_t>(BR_TRANSACTION):
         {
             binder_transaction_data tr;
             result = mIn.read(&tr, sizeof(tr));
@@ -1109,7 +1109,7 @@ status_t IPCThreadState::executeCommand(int32_t cmd)
         }
         break;
     
-    case BR_DEAD_BINDER:
+    case static_cast<int32_t>(BR_DEAD_BINDER):
         {
             BpBinder *proxy = (BpBinder*)mIn.readInt32();
             proxy->sendObituary();
@@ -1117,7 +1117,7 @@ status_t IPCThreadState::executeCommand(int32_t cmd)
             mOut.writeInt32((int32_t)proxy);
         } break;
         
-    case BR_CLEAR_DEATH_NOTIFICATION_DONE:
+    case static_cast<int32_t>(BR_CLEAR_DEATH_NOTIFICATION_DONE):
         {
             BpBinder *proxy = (BpBinder*)mIn.readInt32();
             proxy->getWeakRefs()->decWeak(proxy);
